requirements.txt
absl-py==1.0.0
alembic==1.13.1
asgiref==3.8.1
astunparse==1.6.3
backports.zoneinfo==0.2.1
black==22.3.0
blinker==1.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.6
cryptography==37.0.2
Deprecated==1.2.13
Django==4.2.11
dlib==19.19.0
Flask==2.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
flatbuffers==2.0
gast==0.4.0
gcloud==0.18.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
greenlet==3.0.3
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
heroku==0.1.4
httplib2==0.20.4
idna==3.3
importlib-metadata==4.10.0
importlib_resources==6.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.3.2
jwcrypto==1.3.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Mako==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.24.4
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==24.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase==3.0.8
Pyrebase4==4.5.0
python-dateutil==1.5
python-jwt==3.3.2
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.29
sqlparse==0.4.4
sseclient==0.0.27
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.11.0
tensorflow-cpu==2.8.1
tensorflow-estimator==2.8.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.4.0
tomli==2.0.1
typing_extensions==4.0.1
tzdata==2024.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
